requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytz>=2023.3
schedule>=1.2.0
